+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Bug fixes:
 * [OLMIS-7574](https://openlmis.atlassian.net/browse/OLMIS-7574): Fixed wrong stock card calculations by adding unaccounted quantity validation
 * [OLMIS-7581](https://openlmis.atlassian.net/browse/OLMIS-7581): Fixed wrong order of displaying products after adding a new one in Physical Inventory - mobile app
 
+Improvements:
+* [OLMIS-7588](https://openlmis.atlassian.net/browse/OLMIS-7588): Physical Inventory validation for unaccounted quantity - mobile app
+
 2.1.2 / 2022-04-21
 ==================